% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUPECcv.R
\name{AUPECcv}
\alias{AUPECcv}
\title{Estimation of the Area Under Prescription Evaluation Curve (AUPEC) in Randomized Experiments Under Cross Validation}
\usage{
AUPECcv(T, tau, Y, ind, centered = TRUE)
}
\arguments{
\item{T}{A vector of the unit-level binary treatment receipt variable for each sample.}

\item{tau}{A matrix where the \code{i}th column is the unit-level continuous score for treatment assignment generated in the \code{i}th fold.}

\item{Y}{The outcome variable of interest.}

\item{ind}{A vector of integers (between 1 and number of folds inclusive) indicating which testing set does each sample belong to.}

\item{centered}{If \code{TRUE}, the outcome variables would be centered before processing. This minimizes
the variance of the estimator. Default is \code{TRUE}.}
}
\value{
A list that contains the following items: \item{aupec}{The estimated
Population Average Prescription Effect.} \item{sd}{The estimated standard deviation
of PAPE.}
}
\description{
This function estimates AUPEC. The details of the methods for this design are given in Imai and Li (2019).
}
\references{
Imai and Li (2019). \dQuote{Experimental Evaluation of Individualized Treatment Rules},
}
\author{
Michael Lingzhi Li, Operations Research Center, Massachusetts Institute of Technology
\email{mlli@mit.edu}, \url{http://mlli.mit.edu};
}
\keyword{evaluation}
