% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_qoi.R
\name{compute_qoi}
\alias{compute_qoi}
\title{Compute Quantities of Interest}
\usage{
compute_qoi(fit_obj, algorithms)
}
\arguments{
\item{fit_obj}{An output object from \code{itr_single_outcome} function.}
}
\description{
Compute Quantities of Interest
}
