% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_qoi.R
\name{compute_qoi_user}
\alias{compute_qoi_user}
\title{Compute Quantities of Interest (PAPE, PAPEp, PAPDp, AUPEC, GATE, GATEcv) with user defined functions}
\usage{
compute_qoi_user(
  user_function,
  algorithms,
  treatment,
  outcome,
  tau,
  data,
  ngates,
  ...
)
}
\arguments{
\item{user_function}{A user-defined function to create an ITR. The function should take the data as input and return an ITR. The output is a vector of the unit-level binary treatment that would have been assigned by the individualized treatment rule. The default is \code{NULL}, which means the ITR will be estimated from the \code{estimate_itr}.}

\item{algorithms}{A vector of the names of the algorithms to be used.}

\item{treatment}{A binary treatment variable in the sample.}

\item{outcome}{A vector of the outcome variable of interest for each sample.}

\item{tau}{A vector of the unit-level continuous score for treatment assignment. We assume those that have tau<0 should not have treatment. Conditional Average Treatment Effect is one possible measure.}

\item{data}{A data frame containing the variables of interest.}

\item{ngates}{The number of gates to be used in the GATE function.}

\item{...}{Additional arguments to be passed to the user-defined function.}
}
\description{
Compute Quantities of Interest (PAPE, PAPEp, PAPDp, AUPEC, GATE, GATEcv) with user defined functions
}
