% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_helpers.R
\name{create_ml_args_bart}
\alias{create_ml_args_bart}
\title{Create arguments for bartMachine}
\usage{
create_ml_args_bart(data)
}
\arguments{
\item{data}{A dataset}
}
\description{
Create arguments for bartMachine
}
