% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_helpers.R
\name{create_ml_args_bartc}
\alias{create_ml_args_bartc}
\title{Create arguments for bartCause}
\usage{
create_ml_args_bartc(data)
}
\arguments{
\item{data}{A dataset}
}
\description{
Create arguments for bartCause
}
