% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_helpers.R
\name{create_ml_args_causalforest}
\alias{create_ml_args_causalforest}
\title{Create arguments for causal forest}
\usage{
create_ml_args_causalforest(data)
}
\arguments{
\item{data}{A dataset}
}
\description{
Create arguments for causal forest
}
