% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_helpers.R
\name{create_ml_args_lasso}
\alias{create_ml_args_lasso}
\title{Create arguments for LASSO}
\usage{
create_ml_args_lasso(data)
}
\arguments{
\item{data}{A dataset}
}
\description{
Create arguments for LASSO
}
