% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_helpers.R
\name{create_ml_args_superLearner}
\alias{create_ml_args_superLearner}
\title{Create arguments for super learner}
\usage{
create_ml_args_superLearner(data)
}
\arguments{
\item{data}{A dataset}
}
\description{
Create arguments for super learner
}
