% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_helpers.R
\name{create_ml_args_svm_cls}
\alias{create_ml_args_svm_cls}
\title{Create arguments for SVM classification}
\usage{
create_ml_args_svm_cls(data)
}
\arguments{
\item{data}{A dataset}
}
\description{
Create arguments for SVM classification
}
