% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_helpers.R
\name{create_ml_arguments}
\alias{create_ml_arguments}
\title{Create arguments for ML algorithms}
\usage{
create_ml_arguments(outcome, treatment, data)
}
\arguments{
\item{outcome}{Outcome of interests}

\item{treatment}{Treatment variable}

\item{data}{A dataset}
}
\description{
Create arguments for ML algorithms
}
