% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het.test.R
\name{het.test}
\alias{het.test}
\title{The Heterogeneity Test for Grouped Average Treatment Effects (GATEs) in Randomized Experiments}
\usage{
het.test(T, tau, Y, ngates = 5, centered = TRUE)
}
\arguments{
\item{T}{Binary treatment indicator (0 or 1).}

\item{tau}{Continuous score vector.}

\item{Y}{Outcome vector.}

\item{ngates}{Number of groups (at least 2).}

\item{centered}{Whether to center outcomes before estimation.}
}
\value{
A list with test statistic \code{stat} and p-value \code{pval}.
}
\description{
Tests whether treatment effects are equal across score groups.
}
\details{
See \code{\link{GATE}} for inference details.
}
\examples{
T = c(1,0,1,0,1,0,1,0)
tau = c(0,0.1,0.2,0.3,0.4,0.5,0.6,0.7)
Y = c(4,5,0,2,4,1,-4,3)
hettestlist <- het.test(T,tau,Y,ngates=2)
hettestlist$stat
hettestlist$pval
}
\references{
Imai and Li (2022). \dQuote{Statistical Inference for Heterogeneous Treatment Effects Discovered by Generic Machine Learning in Randomized Experiments},
}
\author{
Michael Lingzhi Li, Technology and Operations Management, Harvard Business School
\email{mili@hbs.edu}, \url{https://www.michaellz.com/};
}
\keyword{evaluation}
