% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{itr_single_outcome}
\alias{itr_single_outcome}
\title{Evaluate ITR for Single Outcome}
\usage{
itr_single_outcome(data, algorithms, params, folds, plim, plot)
}
\description{
Evaluate ITR for Single Outcome
}
