% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAVcv.R
\name{PAVcv}
\alias{PAVcv}
\title{Estimation of the Population Average Value in Randomized Experiments Under Cross Validation}
\usage{
PAVcv(T, That, Y, ind, centered = TRUE)
}
\arguments{
\item{T}{A vector of the unit-level binary treatment receipt variable for each sample.}

\item{That}{A matrix where the \code{i}th column is the unit-level binary treatment that would have been assigned by the
individualized treatment rule generated in the \code{i}th fold. If \code{plim} is specified, please ensure
that the percentage of treatment units of That is lower than the budget constraint.}

\item{Y}{The outcome variable of interest.}

\item{ind}{A vector of integers (between 1 and number of folds inclusive) indicating which testing set does each sample belong to.}

\item{centered}{If \code{TRUE}, the outcome variables would be centered before processing. This minimizes
the variance of the estimator. Default is \code{TRUE}.}
}
\value{
A list that contains the following items: \item{pav}{The estimated
Population Average Value.} \item{sd}{The estimated standard deviation
of PAV.}
}
\description{
This function estimates the Population Average Value. The details of the methods for this design are given in Imai and Li (2019).
}
\examples{
T = c(1,0,1,0,1,0,1,0)
That = matrix(c(0,1,1,0,0,1,1,0,1,0,0,1,1,0,0,1), nrow = 8, ncol = 2)
Y = c(4,5,0,2,4,1,-4,3)
ind = c(rep(1,4),rep(2,4))
pavlist <- PAVcv(T, That, Y, ind)
pavlist$pav
pavlist$sd
}
\references{
Imai and Li (2019). \dQuote{Experimental Evaluation of Individualized Treatment Rules},
}
\author{
Michael Lingzhi Li, Operations Research Center, Massachusetts Institute of Technology
\email{mlli@mit.edu}, \url{http://mlli.mit.edu};
}
\keyword{evaluation}
