% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_plot.R
\name{plot.aupec}
\alias{plot.aupec}
\title{Plot the AUPEC curve}
\usage{
\method{plot}{aupec}(x, Y, T, ...)
}
\arguments{
\item{x}{An object of \code{AUPEC()} class. This is typically an output of \code{AUPEC()} function.}

\item{Y}{The outcome variable.}

\item{T}{The treatment variable.}

\item{...}{Further arguments passed to the function.}
}
\value{
A plot of ggplot2 object. The plot shows the AUPEC curve across each possible budget point for the dataset. Each step increases the budget by 1/n where n is the number of data points.
}
\description{
Plot the AUPEC curve
}
