% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_plot_aupec.R
\name{plot.itr}
\alias{plot.itr}
\title{Plot the AUPEC curve}
\usage{
\method{plot}{itr}(x, outcome = TRUE, ...)
}
\arguments{
\item{x}{An object of \code{estimate_itr} class. This is typically an output of \code{estimate_itr()} function.}

\item{outcome}{Outcome variable. Can be different from the input of \code{run_itr}}

\item{...}{Further arguments passed to the function.}
}
\value{
A plot of ggplot2 object.
}
\description{
Plot the AUPEC curve
}
