% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_plot_aupec.R
\name{plot_aupec}
\alias{plot_aupec}
\title{Plot the AUPEC curve}
\usage{
plot_aupec(fit, data, treatment, outcome, algorithms)
}
\arguments{
\item{fit}{A list that contains the following items:
* `aupec`: The estimated Area Under Prescription Evaluation Curve
* `sd`: The estimated standard deviation of AUPEC
* `aupec_cv`: The estimated Area Under Prescription Evaluation Curve using cross-validation}

\item{data}{The dataset.}

\item{treatment}{A vector of the unit-level binary treatment receipt variable for each sample.}

\item{outcome}{A vector of the outcome variable of interest for each sample.}

\item{algorithms}{A vector of machine learning algorithms.}
}
\description{
Plot the AUPEC curve
}
