% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_plot.R
\name{plot_CI}
\alias{plot_CI}
\title{Plot the uniform confidence interval}
\usage{
plot_CI(x, ...)
}
\arguments{
\item{x}{An object of \code{evaluate_itr()} class. This is typically an output of \code{evaluate_itr()} function.}

\item{...}{Further arguments passed to the function.}
}
\value{
A plot of ggplot2 object.
}
\description{
Plot the uniform confidence interval
}
