% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_plot.R
\name{plot_estimate}
\alias{plot_estimate}
\title{Plot the GATE estimate}
\usage{
plot_estimate(x, type, ...)
}
\arguments{
\item{x}{An table object. This is typically an output of \code{evaluate_itr()} function.}

\item{type}{The metric that you wish to plot. One of GATE, PAPE, PAPEp, or PAPDp.}

\item{...}{Further arguments passed to the function.}
}
\value{
A plot of ggplot2 object.
}
\description{
Plot the GATE estimate
}
