% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_summary.R
\name{print.summary.itr}
\alias{print.summary.itr}
\title{Print}
\usage{
\method{print}{summary.itr}(x, ...)
}
\arguments{
\item{x}{An object of \code{summary.itr} class. This is typically an output of \code{summary.itr()} function.}

\item{...}{Other parameters. Currently not supported.}
}
\description{
Print
}
