% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_summary.R
\name{print.summary.test_itr}
\alias{print.summary.test_itr}
\title{Print}
\usage{
\method{print}{summary.test_itr}(x, ...)
}
\arguments{
\item{x}{An object of \code{summary.test_itr} class. This is typically an output of \code{summary.test_itr()} function.}

\item{...}{Other parameters.}
}
\description{
Print
}
