% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{run_itr}
\alias{run_itr}
\title{Evaluate ITR}
\usage{
run_itr(
  var_outcome,
  var_treatment,
  var_covariates,
  data,
  algorithms,
  plim,
  n_folds,
  plot = FALSE
)
}
\arguments{
\item{var_outcome}{XXX}

\item{var_treatment}{YYY}

\item{data}{A data frame that contains \code{var_outcome} and \code{var_treatment}.}

\item{algorithms}{Machine learning algorithms.}

\item{plim}{Proportion of treated units.}

\item{n_folds}{Number of cross-validation folds.}

\item{options}{List of options.}
}
\description{
Evaluate ITR
}
