% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{star}
\alias{star}
\title{Tennessee’s Student/Teacher Achievement Ratio (STAR) project}
\format{
A data frame with 1911 observations and 14 variables:
\describe{
  \item{treatment}{A binary treatment indicating whether a student is assigned to small class and regular class without an aid}
  \item{g3tlangss}{A continous variable measuring student's writing scores}
  \item{g3treadss}{A continous variable measuring student's reading scores}
  \item{g3tmathss}{A continous variable measuring student's math scores}
\item{gender}{Students' gender} 
\item{race}{Students' race} 
\item{birthmonth}{Students' birth month}
\item{birthyear}{Students' birth year}
\item{SCHLURBN}{Urban or rural} 
\item{GKENRMNT}{Enrollment size} 
\item{GRDRANGE}{Grade range} 
\item{GKFRLNCH}{Number of students on free lunch} 
\item{GKBUSED}{Number of students on school buses} 
\item{GKWHITE}{Percentage of white students}
}
}
\usage{
star
}
\description{
A longitudinal study experimentally evaluating the impacts of class size in early education on various outcomes (Mosteller, 1995)
}
\keyword{datasets}
