% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_summary.R
\name{summary.itr}
\alias{summary.itr}
\title{Summarize estimate_itr output}
\usage{
\method{summary}{itr}(object, ...)
}
\arguments{
\item{object}{An object of \code{estimate_itr} class (typically an output of \code{estimate_itr()} function).}

\item{...}{Other parameters. Currently not supported}
}
\description{
Summarize estimate_itr output
}
