% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_summary.R
\name{summary.test_itr}
\alias{summary.test_itr}
\title{Summarize test_itr output}
\usage{
\method{summary}{test_itr}(object, ...)
}
\arguments{
\item{object}{An object of \code{test_itr} class (typically an output of \code{test_itr()} function).}

\item{...}{Other parameters.}
}
\description{
Summarize test_itr output
}
