% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_helpers.R
\name{create_ml_args_rf_cls}
\alias{create_ml_args_rf_cls}
\title{Create argments for random forest for classification}
\usage{
create_ml_args_rf_cls(data)
}
\arguments{
\item{data}{A dataset}
}
\description{
Create argments for random forest for classification
}
