% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{estimate_itr}
\alias{estimate_itr}
\title{Estimate quantity of interests}
\usage{
estimate_itr(fit, algorithms, outcome)
}
\arguments{
\item{fit}{Fitted model. Usually an output from \code{run_itr}}

\item{algorithms}{Machine learning algorithms}

\item{outcome}{Outcome variable. Need to be the same as the input of \code{run_itr}}
}
\value{
An object of \code{itr} class
}
\description{
Estimate quantity of interests
}
