% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{fit_itr}
\alias{fit_itr}
\title{Estimate ITR for Single Outcome}
\usage{
fit_itr(
  data,
  algorithms,
  params,
  folds,
  budget,
  user_model,
  threshold,
  meta_learner,
  ...
)
}
\arguments{
\item{data}{A dataset.}

\item{algorithms}{Machine learning algorithms.}

\item{params}{A list of parameters.}

\item{folds}{Number of folds.}

\item{budget}{The maximum percentage of population that can be treated under the budget constraint.}

\item{user_model}{User's own function to estimated the ITR.}

\item{...}{Additional arguments passed to \code{caret::train}}
}
\value{
A list of estimates.
}
\description{
Estimate ITR for Single Outcome
}
