% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAV.R
\name{PAV}
\alias{PAV}
\title{Estimation of the Population Average Value in Randomized Experiments}
\usage{
PAV(T, That, Y, centered = TRUE)
}
\arguments{
\item{T}{A vector of the unit-level binary treatment receipt variable for each sample.}

\item{That}{A vector of the unit-level binary treatment that would have been assigned by the
individualized treatment rule. If \code{budget} is specified, please ensure
that the percentage of treatment units of That is lower than the budget constraint.}

\item{Y}{A vector of the outcome variable of interest for each sample.}

\item{centered}{If \code{TRUE}, the outcome variables would be centered before processing. This minimizes
the variance of the estimator. Default is \code{TRUE}.}
}
\value{
A list that contains the following items: \item{pav}{The estimated
Population Average Value.} \item{sd}{The estimated standard deviation
of PAV.}
}
\description{
This function estimates the Population Average Value. The details of the methods for this design are given in Imai and Li (2019).
}
\examples{
T = c(1,0,1,0,1,0,1,0)
That = c(0,1,1,0,0,1,1,0)
Y = c(4,5,0,2,4,1,-4,3)
pavlist <- PAV(T,That,Y)
pavlist$pav
pavlist$sd
}
\references{
Imai and Li (2019). \dQuote{Experimental Evaluation of Individualized Treatment Rules},
}
\author{
Michael Lingzhi Li, Technology and Operations Management, Harvard Business School
\email{mili@hbs.edu}, \url{https://www.michaellz.com/};
}
\keyword{evaluation}
