% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_plot.R
\name{plot.aupec_cv}
\alias{plot.aupec_cv}
\title{Plot the AUPEC curve under cross-validation}
\usage{
\method{plot}{aupec_cv}(x, tau, tau_cv, Y, T, ind, ...)
}
\arguments{
\item{x}{An object of \code{AUPECcv()} class. This is typically an output of \code{AUPECcv()} function.}

\item{tau}{A vector of the unit-level continuous score for treatment assignment.}

\item{tau_cv}{A matrix where the \code{i}th column is the unit-level continuous score for treatment assignment generated in the \code{i}th fold.}

\item{Y}{The outcome variable.}

\item{T}{The treatment variable.}

\item{ind}{A vector of integers (between 1 and number of folds inclusive) indicating which testing set does each sample belong to.}

\item{...}{Further arguments passed to the function.}
}
\value{
A plot of ggplot2 object. The plot shows the AUPEC curve across each possible budget point for the dataset. Each step increases the budget by 1/n where n is the number of data points.
}
\description{
Plot the AUPEC curve under cross-validation
}
