% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_plot_aupec.R
\name{plot.itr}
\alias{plot.itr}
\title{Plot the AUPEC curve}
\usage{
\method{plot}{itr}(x, ...)
}
\arguments{
\item{x}{An object of \code{evaluate_itr()} class. This is typically an output of \code{evaluate_itr()} function.}

\item{...}{Further arguments passed to the function.}
}
\value{
A plot of ggplot2 object.
}
\description{
Plot the AUPEC curve
}
