% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_plot_aupec.R
\name{plot.itr}
\alias{plot.itr}
\title{Plot the AUPEC curve}
\usage{
\method{plot}{itr}(x, data, treatment, outcome, algorithms, ...)
}
\arguments{
\item{x}{An object of \code{run_itr} class. This is typically an output of \code{run_itr()} function.}

\item{data}{The dataset.}

\item{treatment}{A vector of the unit-level binary treatment receipt variable for each sample.}

\item{outcome}{A vector of the outcome variable of interest for each sample.}

\item{algorithms}{A vector of machine learning algorithms.}

\item{...}{Other parameters. Currently not supported}
}
\value{
A plot of ggplot2 object.
}
\description{
Plot the AUPEC curve
}
