% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{run_itr}
\alias{run_itr}
\title{Evaluate ITR}
\usage{
run_itr(
  outcome,
  treatment,
  covariates,
  data,
  algorithms,
  plim,
  n_folds,
  plot = FALSE
)
}
\arguments{
\item{outcome}{XXX}

\item{treatment}{YYY}

\item{data}{A data frame that contains \code{outcome} and \code{treatment}.}

\item{algorithms}{Machine learning algorithms.}

\item{plim}{Proportion of treated units.}

\item{n_folds}{Number of cross-validation folds.}

\item{options}{List of options.}
}
\description{
Evaluate ITR
}
