% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{run_itr}
\alias{run_itr}
\title{Evaluate ITR}
\usage{
run_itr(outcome, treatment, covariates, data, algorithms, plim, n_folds, ratio)
}
\arguments{
\item{outcome}{Outcome variable (or a list of outcome variables). Only takes in numeric values for both continous outcomes and binary outcomes (0 or 1).}

\item{treatment}{Treatment variable}

\item{covariates}{Covariates included in the model.}

\item{data}{A data frame that contains \code{outcome} and \code{treatment}.}

\item{algorithms}{List of machine learning algorithms.}

\item{plim}{Proportion of treated units.}

\item{n_folds}{Number of cross-validation folds.}

\item{ratio}{Split ratio between train and test set under sample splitting.}
}
\value{
An object of \code{itr} class
}
\description{
Evaluate ITR
}
