% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_summary.R
\name{summary.itr}
\alias{summary.itr}
\title{Summarize run_itr output}
\usage{
\method{summary}{itr}(object, m, type, ...)
}
\arguments{
\item{object}{An object of \code{run_itr} class.}

\item{m}{A numeric value indicates which outcome to look at from the vector of \code{outcomes}}

\item{type}{A string indicates the quantity of interests to compute. \code{type} includes PAPE, PAPEp, PAPDp}

\item{...}{Other parameters. Currently not supported}
}
\description{
Summarize run_itr output
}
