% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_summary.R
\name{summary.itr}
\alias{summary.itr}
\title{Summarize run_itr output}
\usage{
\method{summary}{itr}(object, m = 1, ...)
}
\arguments{
\item{object}{An object of \code{run_itr} class}

\item{m}{A numeric value indicates which element to focus on from the vector of \code{outcomes}. The default is the first outcome in the vector \code{outcomes}.}

\item{...}{Other parameters. Currently not supported}
}
\description{
Summarize run_itr output
}
