% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itr_summary.R
\name{summary.itr}
\alias{summary.itr}
\title{Summarize run_itr output}
\usage{
\method{summary}{itr}(object, outcome = TRUE, ...)
}
\arguments{
\item{object}{An object of \code{run_itr} class}

\item{outcome}{Outcome variable. Can be different from the input of \code{run_itr}}

\item{...}{Other parameters. Currently not supported}
}
\description{
Summarize run_itr output
}
