% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{test_itr}
\alias{test_itr}
\title{Conduct hypothesis tests}
\usage{
test_itr(fit, nsim = 10000, ...)
}
\arguments{
\item{fit}{Fitted model. Usually an output from \code{estimate_itr}}

\item{nsim}{Number of Monte Carlo simulations used to simulate the null distributions. Default is 10000.}

\item{...}{Further arguments passed to the function.}

\item{ngates}{The number of groups to separate the data into. The groups are determined by \code{tau}. Default is 5.}
}
\value{
An object of \code{itr} class
}
\description{
Conduct hypothesis tests
}
