% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{test_itr}
\alias{test_itr}
\title{Conduct hypothesis tests}
\usage{
test_itr(fit, nsim = 1000, ...)
}
\arguments{
\item{fit}{Fitted model. Usually an output from \code{estimate_itr}}

\item{nsim}{Number of Monte Carlo simulations used to simulate the null distributions. Default is 1000.}

\item{...}{Further arguments passed to the function.}
}
\value{
An object of \code{itr} class
}
\description{
Conduct hypothesis tests
}
